no message
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@96337 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0faa57327b9ae3fc2b8b18b3c0fa5fbca12823ac

+++ Online/Online/CCPCHist/CCPCHist/CCPCHist.h
@@ -93,6 +93,6 @@ public:
 	int fill(bintype x, bintype y);
 	int getsums (int *ne, bintype *to) ;
   int getsum2s(int *ne, bintype *to );
-  int getentries(int *ne, int *to);
+  int getentries(int *ne, float *to);
 };
 #endif
